test Directory.get_audio_file/1 for published and unpublished episodes
Still working out on how audio file associations are best treated. Seems like it's clearer if we have actual attachment schemas instead of just an abstract one.

Can't figure out how to combine ex_machina with uploads so I added a helper function in the test.

Wondering about the "protocol" of how Directory.* methods return values. I suggest {:ok, value} and {:error, reason} while the reason is specific and the controller decides what if it wants to pass that back to the user. We might not want to have different public responses for a file that does not exist vs. an unpublished file.