Separate public and authenticated data access (#76)
There are now two GraphQL schemas, one for public and one for protected access. Main intention is a clear separation in code, it's still to be seen if it is practical from a GraphQL client point of view.

There are still some TODOs in this restructuring but it's time to bring the changes back to master.