fix(api/contribution): audio permission check
This logic existed before in a similar way but was removed[1]. I'm not sure why.
Also, we had a discussion[2] before on checking permissions on objects that don't have permissions and decided to avoid that and instead pass in the proper parent object into permission checks. However in that discussion the topic was chapters and it was trivial to just pass in the parent instead. Checking the permission on the Audio however is more envolved and I want this logic in _one_ place only.

[1]: https://github.com/podlove/radiator/commit/9b878fc7#diff-8fce5c34154f97d93f212e64bd95fee7L61
[2]: https://github.com/podlove/radiator/pull/152#discussion_r300455042